Moved oidc functionality to actions/core
