+++ packages/core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@actions/core",
-  "version": "1.4.0",
+  "version": "1.4.1",
   "description": "Actions core lib",
   "keywords": [
     "github",
@@ -36,6 +36,7 @@
     "url": "https://github.com/actions/toolkit/issues"
   },
   "devDependencies": {
+    "@actions/http-client": "^1.0.11",
     "@types/node": "^12.0.2"
   }
 }
