+++ packages/core/src/core.ts
@@ -5,6 +5,8 @@ import {toCommandValue} from './utils'
 import * as os from 'os'
 import * as path from 'path'
 
+import {getIDTokenUrl, parseJson, postCall} from './oidc-utils'
+
 /**
  * Interface for getInput options
  */
@@ -283,4 +285,21 @@ export function saveState(name: string, value: any): void {
  */
 export function getState(name: string): string {
   return process.env[`STATE_${name}`] || ''
+}
+
+export async function getIDToken(audience: string): Promise<string> {
+  try {
+    // New ID Token is requested from action service
+    let id_token_url: string = getIDTokenUrl()
+
+    debug(`ID token url is ${id_token_url}`)
+
+    let body: string = await postCall(id_token_url, audience)
+    let id_token = parseJson(body)
+    return id_token
+
+  } catch (error) {
+    setFailed(error.message)
+    return error.message
+  }
 }
