--- packages/oidc-client/package.json
@@ -1,55 +1,0 @@
-{
-  "name": "@actions/oidc-client",
-  "version": "0.0.0",
-  "description": "To get id token from oidc provider",
-  "main": "lib/main.js",
-  "types": "lib/main.d.ts",
-  "scripts": {
-    "build": "tsc",
-    "format": "prettier --write **/*.ts",
-    "format-check": "prettier --check **/*.ts",
-    "package": "ncc build --source-map --license licenses.txt",
-    "test": "jest",
-    "all": "npm run build && npm run format && npm run package && npm test",
-    "tsc": "tsc"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/actions/toolkit.git",
-    "directory": "packages/oidc-client"
-  },
-  "keywords": [
-    "actions",
-    "node",
-    "setup"
-  ],
-  "directories": {
-    "lib": "lib"
-  },
-  "license": "MIT",
-  "dependencies": {
-    "@actions/core": "^1.2.6",
-    "@actions/http-client": "^1.0.11",
-    "@octokit/core": "^3.4.0",
-    "jwt-decode": "3.1.2"
-  },
-  "devDependencies": {
-    "@types/jest": "^26.0.15",
-    "@types/node": "^14.14.9",
-    "@vercel/ncc": "^0.25.1",
-    "jest": "^26.6.3",
-    "jest-circus": "^26.6.3",
-    "js-yaml": "^3.14.0",
-    "prettier": "2.2.1",
-    "ts-jest": "^26.4.4",
-    "typescript": "^4.1.3",
-    "jwt-decode": "3.1.2"
-  },
-  "bugs": {
-    "url": "https://github.com/actions/toolkit/issues"
-  },
-  "homepage": "https://github.com/actions/toolkit#readme"
-}
